requirements.txt
Django>=3.2.4
djangorestframework>=3.12.4
django_extensions>=3.1.3
yahoo-fin>=0.8.8
yfinance>=0.1.59
APScheduler>=3.7.0
